pyproject.toml
[tool.poetry]
name = "snekcord"
version = "1.0.0a"
description = "A work-in-progress Discord API wrapper written in Python."
authors = ["asleep-cult <asleep.cult@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
wsaio = "^1.0.0"
httpx = "^0.21.3"
loguru = "^0.5.3"
attrs = "^21.4.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.16.0"
blue = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.blue]
line-length = 100

[tool.pyright]
reportUnnecessaryIsInstance = false
